+++ CHANGELOG.md
@@ -3,7 +3,13 @@
 
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased](https://github.com/uber/nebula.gl/compare/v0.6.1...master) -->
+## [Unreleased](https://github.com/uber/nebula.gl/compare/v0.6.1...master)
+
+### Added
+
+### Changed
+
+* `EditableGeoJsonLayer`'s `drawAtFront` prop should now be supplied via `modeConfig` prop
 
 ## [0.6.1](https://github.com/uber/nebula.gl/compare/v0.5.1...v0.6.1) - 2018-10-10